helpers.js file has helper func to generate random messages, call in discussion and get-debate api routes
